Testing: Move MockDossierTypes to og.testing.helpers.
